+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/features/one_tap/OneTap.kt
@@ -1,5 +1,6 @@
 package com.mercadopago.android.px.internal.features.one_tap
 
+import android.net.Uri
 import com.mercadolibre.android.cardform.internal.LifecycleListener
 import com.mercadopago.android.px.internal.base.MvpView
 import com.mercadopago.android.px.internal.features.one_tap.slider.HubAdapter
@@ -7,7 +8,9 @@ import com.mercadopago.android.px.internal.viewmodel.drawables.DrawableFragmentI
 import com.mercadopago.android.px.internal.viewmodel.SplitSelectionState
 import com.mercadopago.android.px.internal.features.one_tap.installments.InstallmentRowHolder
 import com.mercadopago.android.px.internal.view.ElementDescriptorView
+import com.mercadopago.android.px.model.DiscountConfigurationModel
 import com.mercadopago.android.px.model.internal.DisabledPaymentMethod
+import com.mercadopago.android.px.model.StatusMetadata
 import com.mercadopago.android.px.core.DynamicDialogCreator
 import com.mercadopago.android.px.internal.experiments.Variant
 import com.mercadopago.android.px.internal.features.generic_modal.GenericDialogItem
@@ -21,8 +24,6 @@ import com.mercadopago.android.px.internal.viewmodel.FlowConfigurationModel
 import com.mercadopago.android.px.internal.features.pay_button.PaymentState
 import com.mercadopago.android.px.model.Currency
 import com.mercadopago.android.px.model.PayerCost
-import com.mercadopago.android.px.model.DiscountConfigurationModel
-import com.mercadopago.android.px.model.StatusMetadata
 import com.mercadopago.android.px.model.internal.Application
 import com.mercadopago.android.px.model.internal.PaymentConfiguration
 import com.mercadopago.android.px.model.one_tap.CheckoutBehaviour
@@ -56,7 +57,7 @@ internal interface OneTap {
         fun showGenericDialog(item: GenericDialogItem)
         fun startAddNewCardFlow(cardFormWrapper: CardFormWrapper?)
         fun startDeepLink(deepLink: String)
-        fun onDeepLinkReceived()
+        fun onDeepLinkReceived(uri: Uri)
         fun showLoading()
         fun hideLoading()
         fun configurePaymentMethodHeader(variant: List<@JvmSuppressWildcards Variant>)
@@ -85,9 +86,11 @@ internal interface OneTap {
         fun handleDeepLink()
         fun onPostPaymentAction(postPaymentAction: PostPaymentAction)
         fun onCardAdded(cardId: String, callback: LifecycleListener.Callback)
-        fun onCardFormResult()
+        fun onBankAccountAdded(uri: Uri)
+        fun onCardAddedResult()
         fun onApplicationChanged(paymentTypeId: String)
         fun onGetViewTrackPath(callback: ConfirmButton.ViewTrackPathCallback)
+        fun resolveDeepLink(uri: Uri)
     }
 
     enum class NavigationState {