+++ px-services/src/main/java/com/mercadopago/android/px/model/OfflinePaymentTypesMetadata.java
@@ -14,7 +14,8 @@ public final class OfflinePaymentTypesMetadata implements Parcelable, Serializab
     private final Text label;
     private final Text description;
     private final List<OfflinePaymentType> paymentTypes;
-    private final DisplayInfo displayInfo;
+    @Nullable private final OfflinePaymentTypeDisplayInfo displayInfo;
+    @Nullable private final GenericCardDisplayInfo genericCardDisplayInfo;
 
     public static final Creator<OfflinePaymentTypesMetadata> CREATOR = new Creator<OfflinePaymentTypesMetadata>() {
         @Override
@@ -44,15 +45,21 @@ public final class OfflinePaymentTypesMetadata implements Parcelable, Serializab
     }
 
     @Nullable
-    public DisplayInfo getDisplayInfo() {
+    public OfflinePaymentTypeDisplayInfo getDisplayInfo() {
         return displayInfo;
+    }
+
+    @Nullable
+    public GenericCardDisplayInfo getGenericCardDisplayInfo() {
+        return genericCardDisplayInfo;
     }
 
     protected OfflinePaymentTypesMetadata(final Parcel in) {
         label = in.readParcelable(Text.class.getClassLoader());
         description = in.readParcelable(Text.class.getClassLoader());
         paymentTypes = in.createTypedArrayList(OfflinePaymentType.CREATOR);
-        displayInfo = in.readParcelable(DisplayInfo.class.getClassLoader());
+        displayInfo = in.readParcelable(OfflinePaymentTypeDisplayInfo.class.getClassLoader());
+        genericCardDisplayInfo = in.readParcelable(GenericCardDisplayInfo.class.getClassLoader());
     }
 
     @Override
@@ -66,41 +73,6 @@ public final class OfflinePaymentTypesMetadata implements Parcelable, Serializab
         dest.writeParcelable(description, flags);
         dest.writeTypedList(paymentTypes);
         dest.writeParcelable(displayInfo, flags);
-    }
-
-    public final static class DisplayInfo implements Parcelable, Serializable {
-
-        private final Text bottomDescription;
-
-        public Text getBottomDescription() {
-            return bottomDescription;
-        }
-
-        public static final Creator<DisplayInfo> CREATOR = new Creator<DisplayInfo>() {
-            @Override
-            public DisplayInfo createFromParcel(final Parcel in) {
-                return new DisplayInfo(in);
-            }
-
-            @Override
-            public DisplayInfo[] newArray(final int size) {
-                return new DisplayInfo[size];
-            }
-        };
-
-        protected DisplayInfo(final Parcel in) {
-            bottomDescription = in.readParcelable(Text.class.getClassLoader());
-        }
-
-        @Override
-        public void writeToParcel(final Parcel dest, final int flags) {
-            dest.writeParcelable(bottomDescription, flags);
-        }
-
-        @Override
-        public int describeContents() {
-            return 0;
-        }
-
+        dest.writeParcelable(genericCardDisplayInfo, flags);
     }
 }