+++ CHANGELOG.md
@@ -1,11 +1,18 @@
+## Unreleased
+_##_05_2022
+* FIX - Adjust track offline_methods
+
+## VERSION 4.114.1
+_05_05_2022
+* FEATURE - Bank account registration.
+* ENHANCEMENT - Update Total value when user selects an installment or split payment or both.
+
 ## VERSION 4.114.0
 _04_05_2022
 * FEATURE - New pending screen to debin status detail.
 * FEATURE - Custom behavior to support custom texts in payment methods.
 * FEATURE - Rules configuration is supported to apply in the means of payment.
 * ENHANCEMENT - New checkout type on tracker added
-* ENHANCEMENT - Update Total value when user selects an installment or split payment or both.
-* FIX - Adjust track offline_methods
 
 ## VERSION 4.113.0
 _28_04_2022_